requirements.txt
fastapi
uvicorn
langchain
openai
python-dotenv
langchain-community